+++ tricorder/CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to the [PVP](https://pvp.haskell.org/).
   build process. See [Configuring Tricorder](/docs/configuring-tricorder.md)
   for more information.
 
+### Fixed
+
+- Unable to correctly parse `packages` lists in `cabal.project` files when the
+  list was formatted as a single-line, comma-separated list.
+
 ## [0.2.1.0] - 2026-08-17
 
 ### Added